+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/_internal/isKotlinTest.kt
@@ -1,0 +1,6 @@
+package com.javiersc.hubdle.project.extensions.kotlin._internal
+
+import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
+
+internal val KotlinSourceSet.isKotlinTest: Boolean
+    get() = name.startsWith("test") || name.endsWith("Test")